+++ src/inmanta/server/server.py
@@ -15,25 +15,18 @@
 
     Contact: code@inmanta.com
 """
-import datetime
 import logging
 import os
 import uuid
-from collections import defaultdict
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Set, Tuple, cast
-from uuid import UUID
+from typing import TYPE_CHECKING, Dict, List, Optional, cast
 
-import asyncpg
 import importlib_metadata
 from tornado import locks
 
-from inmanta import const, data
-from inmanta.const import STATE_UPDATE, TERMINAL_STATES, TRANSIENT_STATES, VALID_STATES_ON_STATE_UPDATE
-from inmanta.data import ResourceVersionIdStr
+from inmanta import data
 from inmanta.data.model import ExtensionStatus, SliceStatus, StatusResponse
 from inmanta.protocol import exceptions, methods
 from inmanta.protocol.common import attach_warnings
-from inmanta.protocol.exceptions import BadRequest, ServerError
 from inmanta.resources import Id
 from inmanta.server import (
     SLICE_AGENT_MANAGER,
@@ -45,94 +38,15 @@ from inmanta.server import (
 )
 from inmanta.server import config as opt
 from inmanta.server import protocol
-from inmanta.types import Apireturn, ArgumentTypes, JsonType, PrimitiveTypes, Warnings
+from inmanta.types import Apireturn, JsonType, Warnings
 
 LOGGER = logging.getLogger(__name__)
 
 if TYPE_CHECKING:
     from inmanta.server.agentmanager import AgentManager
-    from inmanta.server.compilerservice import CompilerService
+    from inmanta.server.services.compilerservice import CompilerService
 
 DBLIMIT = 100000
-
-
-def error_and_log(message: str, **context: Any) -> None:
-    """
-    :param message: message to return both to logger and to remote caller
-    :param context: additional context to attach to log
-    """
-    ctx = ",".join([f"{k}: {v}" for k, v in context.items()])
-    LOGGER.error("%s %s", message, ctx)
-    raise BadRequest(message)
-
-
-class ResourceActionLogLine(logging.LogRecord):
-    """ A special log record that is used to report log lines that come from the agent
-    """
-
-    def __init__(self, logger_name: str, level: str, msg: str, created: datetime.datetime) -> None:
-        super().__init__(
-            name=logger_name,
-            level=level,
-            pathname="(unknown file)",
-            lineno=0,
-            msg=msg,
-            args=[],
-            exc_info=None,
-            func=None,
-            sinfo=None,
-        )
-
-        self.created = created.timestamp()
-        self.msecs = (self.created - int(self.created)) * 1000
-        self.relativeCreated = (self.created - logging._startTime) * 1000
-
-
-class DatabaseSlice(protocol.ServerSlice):
-    """Slice to initialize the database"""
-
-    def __init__(self) -> None:
-        super(DatabaseSlice, self).__init__(SLICE_DATABASE)
-        self._pool: Optional[asyncpg.pool.Pool] = None
-
-    async def start(self) -> None:
-        await self.connect_database()
-
-    async def stop(self) -> None:
-        await self.disconnect_database()
-        self._pool = None
-
-    def get_dependencies(self) -> List[str]:
-        return []
-
-    async def connect_database(self) -> None:
-        """ Connect to the database
-        """
-        database_host = opt.db_host.get()
-        database_port = opt.db_port.get()
-
-        database_username = opt.db_username.get()
-        database_password = opt.db_password.get()
-        self._pool = await data.connect(database_host, database_port, opt.db_name.get(), database_username, database_password)
-        LOGGER.info("Connected to PostgreSQL database %s on %s:%d", opt.db_name.get(), database_host, database_port)
-
-    async def disconnect_database(self) -> None:
-        """ Disconnect the database
-        """
-        await data.disconnect()
-
-    async def get_status(self) -> Dict[str, ArgumentTypes]:
-        """ Get the status of the database connection
-        """
-        status = {"connected": self._pool is not None, "database": opt.db_name.get(), "host": opt.db_host.get()}
-
-        if self._pool is not None:
-            status["max_pool"] = self._pool._maxsize
-            status["open_connections"] = (
-                len([x for x in self._pool._holders if x._con is not None and not x._con.is_closed()]),
-            )
-
-        return status
 
 
 class Server(protocol.ServerSlice):
@@ -152,13 +66,6 @@ class Server(protocol.ServerSlice):
         self.setup_dashboard()
         self.dryrun_lock = locks.Lock()
 
-        self._resource_action_loggers: Dict[uuid.UUID, logging.Logger] = {}
-        self._resource_action_handlers: Dict[uuid.UUID, logging.Handler] = {}
-
-        self._increment_cache: Dict[uuid.UUID, Tuple[Set[ResourceVersionIdStr], List[ResourceVersionIdStr]]] = {}
-        # lock to ensure only one inflight request
-        self._increment_cache_locks: Dict[uuid.UUID, locks.Lock] = defaultdict(lambda: locks.Lock())
-
     def get_dependencies(self) -> List[str]:
         return [SLICE_SESSION_MANAGER, SLICE_DATABASE]
 
@@ -170,92 +77,6 @@ class Server(protocol.ServerSlice):
         self._server_storage: Dict[str, str] = self.check_storage()
         self.agentmanager: "AgentManager" = cast("AgentManager", server.get_slice(SLICE_AGENT_MANAGER))
         self.compiler: "CompilerService" = cast("CompilerService", server.get_slice(SLICE_COMPILER))
-
-    async def start(self) -> None:
-        self.schedule(self._purge_versions, opt.server_purge_version_interval.get())
-        self.schedule(data.ResourceAction.purge_logs, opt.server_purge_resource_action_logs_interval.get())
-
-        self.add_background_task(self._purge_versions())
-
-        await super().start()
-
-    async def stop(self) -> None:
-        await super().stop()
-        self._close_resource_action_loggers()
-
-    @staticmethod
-    def get_resource_action_log_file(environment: uuid.UUID) -> str:
-        """Get the correct filename for the given environment
-        :param environment: The environment id to get the file for
-        :return: The path to the logfile
-        """
-        return os.path.join(opt.log_dir.get(), opt.server_resource_action_log_prefix.get() + str(environment) + ".log")
-
-    def get_resource_action_logger(self, environment: uuid.UUID) -> logging.Logger:
-        """Get the resource action logger for the given environment. If the logger was not created, create it.
-        :param environment: The environment to get a logger for
-        :return: The logger for the given environment.
-        """
-        if environment in self._resource_action_loggers:
-            return self._resource_action_loggers[environment]
-
-        resource_action_log = self.get_resource_action_log_file(environment)
-
-        file_handler = logging.handlers.WatchedFileHandler(filename=resource_action_log, mode="a+")
-        # Most logs will come from agents. We need to use their level and timestamp and their formatted message
-        file_handler.setFormatter(logging.Formatter(fmt="%(message)s"))
-        file_handler.setLevel(logging.DEBUG)
-
-        resource_action_logger = logging.getLogger(const.NAME_RESOURCE_ACTION_LOGGER).getChild(str(environment))
-        resource_action_logger.setLevel(logging.DEBUG)
-        resource_action_logger.addHandler(file_handler)
-
-        self._resource_action_loggers[environment] = resource_action_logger
-        self._resource_action_handlers[environment] = file_handler
-
-        return resource_action_logger
-
-    def _close_resource_action_loggers(self) -> None:
-        """Close all resource action loggers and their associated handlers"""
-        try:
-            while True:
-                env, logger = self._resource_action_loggers.popitem()
-                self.close_resource_action_logger(env, logger)
-        except KeyError:
-            pass
-
-    def close_resource_action_logger(self, env: uuid.UUID, logger: logging.Logger = None) -> None:
-        """Close the given logger for the given env.
-        :param env: The environment to close the logger for
-        :param logger: The logger to close, if the logger is none it is retrieved
-        """
-        if logger is None:
-            if env in self._resource_action_loggers:
-                logger = self._resource_action_loggers.pop(env)
-            else:
-                return
-
-        handler = self._resource_action_handlers.pop(env)
-        logger.removeHandler(handler)
-        handler.flush()
-        handler.close()
-
-    def log_resource_action(
-        self, env: uuid.UUID, resource_ids: List[str], log_level: int, ts: datetime.datetime, message: str
-    ) -> None:
-        """Write the given log to the correct resource action logger"""
-        logger = self.get_resource_action_logger(env)
-        if len(resource_ids) == 0:
-            message = "no resources: " + message
-        elif len(resource_ids) > 1:
-            message = "multiple resources: " + message
-        else:
-            message = resource_ids[0] + ": " + message
-        log_record = ResourceActionLogLine(logger.name, log_level, message, ts)
-        logger.handle(log_record)
-
-    def get_agent_client(self, tid: UUID, endpoint: str) -> Optional[protocol.ReturnClient]:
-        return self.agentmanager.get_agent_client(tid, endpoint)
 
     def setup_dashboard(self) -> None:
         """
@@ -293,29 +114,6 @@ angular.module('inmantaApi.config', []).constant('inmantaConfig', {
         self.add_static_content("/dashboard/config.js", content=content)
         self.add_static_handler("/dashboard", dashboard_path, start=True)
 
-    def clear_env_cache(self, env: data.Environment) -> None:
-        LOGGER.log(const.LOG_LEVEL_TRACE, "Clearing cache for %s", env.id)
-        self._increment_cache[env.id] = None
-
-    async def _purge_versions(self) -> None:
-        """
-            Purge versions from the database
-        """
-        # TODO: move to data and use queries for delete
-        envs = await data.Environment.get_list()
-        for env_item in envs:
-            # get available versions
-            n_versions = opt.server_version_to_keep.get()
-            versions = await data.ConfigurationModel.get_list(environment=env_item.id)
-            if len(versions) > n_versions:
-                LOGGER.info("Removing %s available versions from environment %s", len(versions) - n_versions, env_item.id)
-                version_dict = {x.version: x for x in versions}
-                delete_list = sorted(version_dict.keys())
-                delete_list = delete_list[:-n_versions]
-
-                for v in delete_list:
-                    await version_dict[v].delete_cascade()
-
     def check_storage(self) -> Dict[str, str]:
         """
             Check if the server storage is configured and ready to use.
@@ -336,547 +134,6 @@ angular.module('inmantaApi.config', []).constant('inmantaConfig', {
         dir_map["logs"] = _ensure_directory_exist(opt.log_dir.get())
         return dir_map
 
-    @protocol.handle(methods.get_resource, resource_id="id", env="tid")
-    async def get_resource(
-        self,
-        env: data.Environment,
-        resource_id: str,
-        logs: bool,
-        status: bool,
-        log_action: const.ResourceAction,
-        log_limit: int,
-    ) -> Apireturn:
-        resv = await data.Resource.get(env.id, resource_id)
-        if resv is None:
-            return 404, {"message": "The resource with the given id does not exist in the given environment"}
-
-        if status is not None and status:
-            return 200, {"status": resv.status}
-
-        actions: List[data.ResourceAction] = []
-        if bool(logs):
-            action_name = None
-            if log_action is not None:
-                action_name = log_action.name
-
-            actions = await data.ResourceAction.get_log(
-                environment=env.id, resource_version_id=resource_id, action=action_name, limit=log_limit
-            )
-
-        return 200, {"resource": resv, "logs": actions}
-
-    @protocol.handle(methods.get_resources_for_agent, env="tid")
-    async def get_resources_for_agent(
-        self, env: data.Environment, agent: str, version: str, sid: uuid.UUID, incremental_deploy: bool
-    ) -> Apireturn:
-        if not self.agentmanager.is_primary(env, sid, agent):
-            return 409, {"message": "This agent is not currently the primary for the endpoint %s (sid: %s)" % (agent, sid)}
-        if incremental_deploy:
-            if version is not None:
-                return 500, {"message": "Cannot request increment for a specific version"}
-            result = await self.get_resource_increment_for_agent(env, agent)
-        else:
-            result = await self.get_all_resources_for_agent(env, agent, version)
-        return result
-
-    async def get_all_resources_for_agent(self, env: data.Environment, agent: str, version: str) -> Apireturn:
-        started = datetime.datetime.now()
-        if version is None:
-            version = await data.ConfigurationModel.get_version_nr_latest_version(env.id)
-            if version is None:
-                return 404, {"message": "No version available"}
-
-        else:
-            exists = await data.ConfigurationModel.version_exists(environment=env.id, version=version)
-            if not exists:
-                return 404, {"message": "The given version does not exist"}
-
-        deploy_model = []
-
-        resources = await data.Resource.get_resources_for_version(env.id, version, agent)
-
-        resource_ids = []
-        for rv in resources:
-            deploy_model.append(rv.to_dict())
-            resource_ids.append(rv.resource_version_id)
-
-        now = datetime.datetime.now()
-
-        log_line = data.LogLine.log(logging.INFO, "Resource version pulled by client for agent %(agent)s state", agent=agent)
-        self.log_resource_action(env.id, resource_ids, logging.INFO, now, log_line.msg)
-        ra = data.ResourceAction(
-            environment=env.id,
-            resource_version_ids=resource_ids,
-            action=const.ResourceAction.pull,
-            action_id=uuid.uuid4(),
-            started=started,
-            finished=now,
-            messages=[log_line],
-        )
-        await ra.insert()
-
-        return 200, {"environment": env.id, "agent": agent, "version": version, "resources": deploy_model}
-
-    async def get_resource_increment_for_agent(self, env: data.Environment, agent: str) -> Apireturn:
-        started = datetime.datetime.now()
-
-        version = await data.ConfigurationModel.get_version_nr_latest_version(env.id)
-        if version is None:
-            return 404, {"message": "No version available"}
-
-        increment = self._increment_cache.get(env.id, None)
-        if increment is None:
-            with (await self._increment_cache_locks[env.id].acquire()):
-                increment = self._increment_cache.get(env.id, None)
-                if increment is None:
-                    increment = await data.ConfigurationModel.get_increment(env.id, version)
-                    self._increment_cache[env.id] = increment
-
-        increment_ids, neg_increment = increment
-
-        # set already done to deployed
-        now = datetime.datetime.now()
-
-        def on_agent(res: ResourceVersionIdStr) -> bool:
-            idr = Id.parse_id(res)
-            return idr.get_agent_name() == agent
-
-        neg_increment = [res_id for res_id in neg_increment if on_agent(res_id)]
-
-        logline = {
-            "level": "INFO",
-            "msg": "Setting deployed due to known good status",
-            "timestamp": now.isoformat(timespec="microseconds"),
-            "args": [],
-        }
-        self.add_background_task(
-            self.resource_action_update(
-                env,
-                neg_increment,
-                action_id=uuid.uuid4(),
-                started=now,
-                finished=now,
-                status=const.ResourceState.deployed,
-                # does this require a different ResourceAction?
-                action=const.ResourceAction.deploy,
-                changes={},
-                messages=[logline],
-                change=const.Change.nochange,
-                send_events=False,
-                keep_increment_cache=True,
-            )
-        )
-
-        resources = await data.Resource.get_resources_for_version(env.id, version, agent)
-
-        deploy_model: List[Dict[str, Any]] = []
-        resource_ids: List[str] = []
-        for rv in resources:
-            if rv.resource_version_id not in increment_ids:
-                continue
-
-            def in_requires(req: ResourceVersionIdStr) -> bool:
-                if req in increment_ids:
-                    return True
-                idr = Id.parse_id(req)
-                return idr.get_agent_name() != agent
-
-            rv.attributes["requires"] = [r for r in rv.attributes["requires"] if in_requires(r)]
-
-            deploy_model.append(rv.to_dict())
-            resource_ids.append(rv.resource_version_id)
-
-        ra = data.ResourceAction(
-            environment=env.id,
-            resource_version_ids=resource_ids,
-            action=const.ResourceAction.pull,
-            action_id=uuid.uuid4(),
-            started=started,
-            finished=now,
-            messages=[
-                data.LogLine.log(logging.INFO, "Resource version pulled by client for agent %(agent)s state", agent=agent)
-            ],
-        )
-        await ra.insert()
-
-        return 200, {"environment": env.id, "agent": agent, "version": version, "resources": deploy_model}
-
-    @protocol.handle(methods.list_versions, env="tid")
-    async def list_version(self, env: data.Environment, start: Optional[int] = None, limit: Optional[int] = None) -> Apireturn:
-        if (start is None and limit is not None) or (limit is None and start is not None):
-            raise ServerError("Start and limit should always be set together.")
-
-        if start is None:
-            start = 0
-            limit = data.DBLIMIT
-
-        models = await data.ConfigurationModel.get_versions(env.id, start, limit)
-        count = len(models)
-
-        d = {"versions": models}
-
-        if start is not None:
-            d["start"] = start
-            d["limit"] = limit
-
-        d["count"] = count
-
-        return 200, d
-
-    @protocol.handle(methods.get_version, version_id="id", env="tid")
-    async def get_version(
-        self,
-        env: data.Environment,
-        version_id: int,
-        include_logs: Optional[bool] = None,
-        log_filter: Optional[str] = None,
-        limit: Optional[int] = 0,
-    ) -> Apireturn:
-        version = await data.ConfigurationModel.get_version(env.id, version_id)
-        if version is None:
-            return 404, {"message": "The given configuration model does not exist yet."}
-
-        resources = await data.Resource.get_resources_for_version(env.id, version_id, no_obj=True)
-        if resources is None:
-            return 404, {"message": "The given configuration model does not exist yet."}
-
-        d = {"model": version}
-
-        # todo: batch get_log into single query?
-        d["resources"] = []
-        for res_dict in resources:
-            if bool(include_logs):
-                res_dict["actions"] = await data.ResourceAction.get_log(
-                    env.id, res_dict["resource_version_id"], log_filter, limit
-                )
-
-            d["resources"].append(res_dict)
-
-        d["unknowns"] = await data.UnknownParameter.get_list(environment=env.id, version=version_id)
-
-        return 200, d
-
-    @protocol.handle(methods.delete_version, version_id="id", env="tid")
-    async def delete_version(self, env, version_id):
-        version = await data.ConfigurationModel.get_version(env.id, version_id)
-        if version is None:
-            return 404, {"message": "The given configuration model does not exist yet."}
-
-        await version.delete_cascade()
-        return 200
-
-    @protocol.handle(methods.put_version, env="tid")
-    async def put_version(
-        self,
-        env: data.Environment,
-        version: int,
-        resources: List[JsonType],
-        resource_state: Dict[str, const.ResourceState],
-        unknowns: List[Dict[str, PrimitiveTypes]],
-        version_info: JsonType,
-    ) -> Apireturn:
-        started = datetime.datetime.now()
-
-        agents = set()
-        # lookup for all RV's, lookup by resource id
-        rv_dict = {}
-        # reverse dependency tree, Resource.provides [:] -- Resource.requires as resource_id
-        provides_tree: Dict[str, List[str]] = defaultdict(lambda: [])
-        # list of all resources which have a cross agent dependency, as a tuple, (dependant,requires)
-        cross_agent_dep = []
-        # list of all resources which are undeployable
-        undeployable: List[data.Resource] = []
-
-        resource_objects = []
-        resource_version_ids = []
-        for res_dict in resources:
-            res_obj = data.Resource.new(env.id, res_dict["id"])
-            if res_obj.resource_id in resource_state:
-                res_obj.status = const.ResourceState[resource_state[res_obj.resource_id]]
-                if res_obj.status in const.UNDEPLOYABLE_STATES:
-                    undeployable.append(res_obj)
-
-            # collect all agents
-            agents.add(res_obj.agent)
-
-            attributes = {}
-            for field, value in res_dict.items():
-                if field != "id":
-                    attributes[field] = value
-
-            res_obj.attributes = attributes
-            resource_objects.append(res_obj)
-            resource_version_ids.append(res_obj.resource_version_id)
-
-            rv_dict[res_obj.resource_id] = res_obj
-
-            # find cross agent dependencies
-            agent = res_obj.agent
-            resc_id = res_obj.resource_id
-            if "requires" not in attributes:
-                LOGGER.warning("Received resource without requires attribute (%s)" % res_obj.resource_id)
-            else:
-                for req in attributes["requires"]:
-                    rid = Id.parse_id(req)
-                    provides_tree[rid.resource_str()].append(resc_id)
-                    if rid.get_agent_name() != agent:
-                        # it is a CAD
-                        cross_agent_dep.append((res_obj, rid))
-
-        # hook up all CADs
-        for f, t in cross_agent_dep:
-            res_obj = rv_dict[t.resource_str()]
-            res_obj.provides.append(f.resource_version_id)
-
-        # detect failed compiles
-        def safe_get(input, key, default):
-            if not isinstance(input, dict):
-                return default
-            if key not in input:
-                return default
-            return input[key]
-
-        metadata = safe_get(version_info, const.EXPORT_META_DATA, {})
-        compile_state = safe_get(metadata, const.META_DATA_COMPILE_STATE, "")
-        failed = compile_state == const.Compilestate.failed.name
-
-        resources_to_purge: List[data.Resource] = []
-        if not failed:
-            # search for deleted resources
-            resources_to_purge = await data.Resource.get_deleted_resources(env.id, version, set(rv_dict.keys()))
-
-            previous_requires = {}
-            for res in resources_to_purge:
-                LOGGER.warning("Purging %s, purged resource based on %s" % (res.resource_id, res.resource_version_id))
-
-                attributes = res.attributes.copy()
-                attributes["purged"] = True
-                attributes["requires"] = []
-                res_obj = data.Resource.new(
-                    env.id, resource_version_id="%s,v=%s" % (res.resource_id, version), attributes=attributes
-                )
-                resource_objects.append(res_obj)
-
-                previous_requires[res_obj.resource_id] = res.attributes["requires"]
-                resource_version_ids.append(res_obj.resource_version_id)
-                agents.add(res_obj.agent)
-                rv_dict[res_obj.resource_id] = res_obj
-
-            # invert dependencies on purges
-            for res_id, requires in previous_requires.items():
-                res_obj = rv_dict[res_id]
-                for require in requires:
-                    req_id = Id.parse_id(require)
-
-                    if req_id.resource_str() in rv_dict:
-                        req_res = rv_dict[req_id.resource_str()]
-
-                        req_res.attributes["requires"].append(res_obj.resource_version_id)
-                        res_obj.provides.append(req_res.resource_version_id)
-
-        undeployable_ids: List[str] = [res.resource_id for res in undeployable]
-        # get skipped for undeployable
-        work = list(undeployable_ids)
-        skippeable: Set[str] = set()
-        while len(work) > 0:
-            current = work.pop()
-            if current in skippeable:
-                continue
-            skippeable.add(current)
-            work.extend(provides_tree[current])
-
-        skip_list = sorted(list(skippeable - set(undeployable_ids)))
-
-        try:
-            cm = data.ConfigurationModel(
-                environment=env.id,
-                version=version,
-                date=datetime.datetime.now(),
-                total=len(resources),
-                version_info=version_info,
-                undeployable=undeployable_ids,
-                skipped_for_undeployable=skip_list,
-            )
-            await cm.insert()
-        except asyncpg.exceptions.UniqueViolationError:
-            return 500, {"message": "The given version is already defined. Versions should be unique."}
-
-        await data.Resource.insert_many(resource_objects)
-        await cm.update_fields(total=cm.total + len(resources_to_purge))
-
-        for uk in unknowns:
-            if "resource" not in uk:
-                uk["resource"] = ""
-
-            if "metadata" not in uk:
-                uk["metadata"] = {}
-
-            up = data.UnknownParameter(
-                resource_id=uk["resource"],
-                name=uk["parameter"],
-                source=uk["source"],
-                environment=env.id,
-                version=version,
-                metadata=uk["metadata"],
-            )
-            await up.insert()
-
-        for agent in agents:
-            await self.agentmanager.ensure_agent_registered(env, agent)
-
-        now = datetime.datetime.now()
-        log_line = data.LogLine.log(logging.INFO, "Successfully stored version %(version)d", version=version)
-        self.log_resource_action(env.id, resource_version_ids, logging.INFO, now, log_line.msg)
-        ra = data.ResourceAction(
-            environment=env.id,
-            resource_version_ids=resource_version_ids,
-            action_id=uuid.uuid4(),
-            action=const.ResourceAction.store,
-            started=started,
-            finished=now,
-            messages=[log_line],
-        )
-        await ra.insert()
-        LOGGER.debug("Successfully stored version %d", version)
-
-        self.clear_env_cache(env)
-
-        auto_deploy = await env.get(data.AUTO_DEPLOY)
-        if auto_deploy:
-            LOGGER.debug("Auto deploying version %d", version)
-            push_on_auto_deploy = await env.get(data.PUSH_ON_AUTO_DEPLOY)
-            agent_trigger_method_on_autodeploy = await env.get(data.AGENT_TRIGGER_METHOD_ON_AUTO_DEPLOY)
-            agent_trigger_method_on_autodeploy = const.AgentTriggerMethod[agent_trigger_method_on_autodeploy]
-            await self.release_version(env, version, push_on_auto_deploy, agent_trigger_method_on_autodeploy)
-
-        return 200
-
-    @protocol.handle(methods.release_version, version_id="id", env="tid")
-    async def release_version(
-        self,
-        env: data.Environment,
-        version_id: int,
-        push: bool,
-        agent_trigger_method: Optional[const.AgentTriggerMethod] = None,
-    ) -> Apireturn:
-        model = await data.ConfigurationModel.get_version(env.id, version_id)
-        if model is None:
-            return 404, {"message": "The request version does not exist."}
-
-        await model.update_fields(released=True, result=const.VersionState.deploying)
-
-        if model.total == 0:
-            await model.mark_done()
-            return 200, {"model": model}
-
-        # Already mark undeployable resources as deployed to create a better UX (change the version counters)
-        undep = await model.get_undeployable()
-        undep = [rid + ",v=%s" % version_id for rid in undep]
-
-        now = datetime.datetime.now()
-
-        # not checking error conditions
-        await self.resource_action_update(
-            env,
-            undep,
-            action_id=uuid.uuid4(),
-            started=now,
-            finished=now,
-            status=const.ResourceState.undefined,
-            action=const.ResourceAction.deploy,
-            changes={},
-            messages=[],
-            change=const.Change.nochange,
-            send_events=False,
-        )
-
-        skippable = await model.get_skipped_for_undeployable()
-        skippable = [rid + ",v=%s" % version_id for rid in skippable]
-        # not checking error conditions
-        await self.resource_action_update(
-            env,
-            skippable,
-            action_id=uuid.uuid4(),
-            started=now,
-            finished=now,
-            status=const.ResourceState.skipped_for_undefined,
-            action=const.ResourceAction.deploy,
-            changes={},
-            messages=[],
-            change=const.Change.nochange,
-            send_events=False,
-        )
-
-        if push:
-            # fetch all resource in this cm and create a list of distinct agents
-            agents = await data.ConfigurationModel.get_agents(env.id, version_id)
-            await self.agentmanager._ensure_agents(env, agents)
-
-            for agent in agents:
-                client = self.get_agent_client(env.id, agent)
-                if client is not None:
-                    if not agent_trigger_method:
-                        # Ensure backward compatibility
-                        incremental_deploy = False
-                    else:
-                        incremental_deploy = agent_trigger_method is const.AgentTriggerMethod.push_incremental_deploy
-                    self.add_background_task(client.trigger(env.id, agent, incremental_deploy))
-                else:
-                    LOGGER.warning("Agent %s from model %s in env %s is not available for a deploy", agent, version_id, env.id)
-
-        return 200, {"model": model}
-
-    @protocol.handle(methods.deploy, env="tid")
-    async def deploy(
-        self,
-        env: data.Environment,
-        agent_trigger_method: const.AgentTriggerMethod = const.AgentTriggerMethod.push_full_deploy,
-        agents: List[str] = None,
-    ) -> Apireturn:
-        warnings = []
-
-        # get latest version
-        version_id = await data.ConfigurationModel.get_version_nr_latest_version(env.id)
-        if version_id is None:
-            return 404, {"message": "No version available"}
-
-        # filter agents
-        allagents = await data.ConfigurationModel.get_agents(env.id, version_id)
-        if agents is not None:
-            required = set(agents)
-            present = set(allagents)
-            allagents = list(required.intersection(present))
-            notfound = required - present
-            if notfound:
-                warnings.append(
-                    "Model version %d does not contain agents named [%s]" % (version_id, ",".join(sorted(list(notfound))))
-                )
-
-        if not allagents:
-            return attach_warnings(404, {"message": "No agent could be reached"}, warnings)
-
-        present = set()
-        absent = set()
-
-        await self.agentmanager._ensure_agents(env, allagents)
-
-        for agent in allagents:
-            client = self.get_agent_client(env.id, agent)
-            if client is not None:
-                incremental_deploy = agent_trigger_method is const.AgentTriggerMethod.push_incremental_deploy
-                self.add_background_task(client.trigger(env.id, agent, incremental_deploy))
-                present.add(agent)
-            else:
-                absent.add(agent)
-
-        if absent:
-            warnings.append("Could not reach agents named [%s]" % ",".join(sorted(list(absent))))
-
-        if not present:
-            return attach_warnings(404, {"message": "No agent could be reached"}, warnings)
-
-        return attach_warnings(200, {"agents": sorted(list(present))}, warnings)
-
     @protocol.handle(methods.dryrun_request, version_id="id", env="tid")
     async def dryrun_request(self, env: data.Environment, version_id: int) -> Apireturn:
         model = await data.ConfigurationModel.get_version(environment=env.id, version=version_id)
@@ -893,7 +150,7 @@ angular.module('inmantaApi.config', []).constant('inmantaConfig', {
         await self.agentmanager._ensure_agents(env, agents)
 
         for agent in agents:
-            client = self.get_agent_client(env.id, agent)
+            client = self.agentmanager.get_agent_client(env.id, agent)
             if client is not None:
                 self.add_background_task(client.do_dryrun(env.id, dryrun.id, agent, version_id))
             else:
@@ -970,174 +227,6 @@ angular.module('inmantaApi.config', []).constant('inmantaConfig', {
         with (await self.dryrun_lock.acquire()):
             payload = {"changes": changes, "id_fields": Id.parse_id(resource).to_dict(), "id": resource}
             await data.DryRun.update_resource(dryrun_id, resource, payload)
-
-        return 200
-
-    @protocol.handle(methods.resource_action_update, env="tid")
-    async def resource_action_update(
-        self,
-        env: data.Environment,
-        resource_ids: List[ResourceVersionIdStr],
-        action_id: uuid.UUID,
-        action: const.ResourceAction,
-        started: datetime.datetime,
-        finished: datetime.datetime,
-        status: const.ResourceState,
-        messages: List[Dict[str, Any]],
-        changes: Dict[str, Any],
-        change: const.Change,
-        send_events: bool,
-        keep_increment_cache: bool = False,
-    ) -> Apireturn:
-        # can update resource state
-        is_resource_state_update = action in STATE_UPDATE
-        # this ra is finishing
-        is_resource_action_finished = finished is not None
-
-        if is_resource_state_update:
-            # if status update, status is required
-            if status is None:
-                error_and_log(
-                    "Cannot perform state update without a status.",
-                    resource_ids=resource_ids,
-                    action=action,
-                    action_id=action_id,
-                )
-            # and needs to be valid
-            if status not in VALID_STATES_ON_STATE_UPDATE:
-                error_and_log(
-                    "Status %s is not valid on action %s" % (status, action),
-                    resource_ids=resource_ids,
-                    action=action,
-                    action_id=action_id,
-                )
-            if status in TRANSIENT_STATES:
-                if not is_resource_action_finished:
-                    pass
-                else:
-                    error_and_log(
-                        "The finished field must not be set for transient states",
-                        status=status,
-                        resource_ids=resource_ids,
-                        action=action,
-                        action_id=action_id,
-                    )
-            else:
-                if is_resource_action_finished:
-                    pass
-                else:
-                    error_and_log(
-                        "The finished field must be set for none transient states",
-                        status=status,
-                        resource_ids=resource_ids,
-                        action=action,
-                        action_id=action_id,
-                    )
-
-        # validate resources
-        resources = await data.Resource.get_resources(env.id, resource_ids)
-        if len(resources) == 0 or (len(resources) != len(resource_ids)):
-            return (
-                404,
-                {
-                    "message": "The resources with the given ids do not exist in the given environment. "
-                    "Only %s of %s resources found." % (len(resources), len(resource_ids))
-                },
-            )
-
-        # validate transitions
-        if is_resource_state_update:
-            # no escape from terminal
-            if any(resource.status != status and resource.status in TERMINAL_STATES for resource in resources):
-                LOGGER.error("Attempting to set undeployable resource to deployable state")
-                raise AssertionError("Attempting to set undeployable resource to deployable state")
-
-        # get instance
-        resource_action = await data.ResourceAction.get(action_id=action_id)
-        if resource_action is None:
-            # new
-            if started is None:
-                return 500, {"message": "A resource action can only be created with a start datetime."}
-
-            resource_action = data.ResourceAction(
-                environment=env.id, resource_version_ids=resource_ids, action_id=action_id, action=action, started=started
-            )
-            await resource_action.insert()
-        else:
-            # existing
-            if resource_action.finished is not None:
-                return (
-                    500,
-                    {
-                        "message": "An resource action can only be updated when it has not been finished yet. This action "
-                        "finished at %s" % resource_action.finished
-                    },
-                )
-
-        if len(messages) > 0:
-            resource_action.add_logs(messages)
-            for msg in messages:
-                # All other data is stored in the database. The msg was already formatted at the client side.
-                self.log_resource_action(
-                    env.id,
-                    resource_ids,
-                    const.LogLevel[msg["level"]].value,
-                    datetime.datetime.strptime(msg["timestamp"], const.TIME_ISOFMT),
-                    msg["msg"],
-                )
-
-        if len(changes) > 0:
-            resource_action.add_changes(changes)
-
-        if status is not None:
-            resource_action.set_field("status", status)
-
-        if change is not None:
-            resource_action.set_field("change", change)
-
-        resource_action.set_field("send_event", send_events)
-
-        if finished is not None:
-            resource_action.set_field("finished", finished)
-
-        await resource_action.save()
-
-        if is_resource_state_update:
-            # transient resource update
-            if not is_resource_action_finished:
-                for res in resources:
-                    await res.update_fields(status=status)
-                if not keep_increment_cache:
-                    self.clear_env_cache(env)
-                return 200
-
-            else:
-                # final resource update
-                if not keep_increment_cache:
-                    self.clear_env_cache(env)
-
-                model_version = None
-                for res in resources:
-                    await res.update_fields(last_deploy=finished, status=status)
-                    model_version = res.model
-
-                    if "purged" in res.attributes and res.attributes["purged"] and status == const.ResourceState.deployed:
-                        await data.Parameter.delete_all(environment=env.id, resource_id=res.resource_id)
-
-                await data.ConfigurationModel.mark_done_if_done(env.id, model_version)
-
-                waiting_agents = set(
-                    [
-                        (Id.parse_id(prov).get_agent_name(), res.resource_version_id)
-                        for res in resources
-                        for prov in res.provides
-                    ]
-                )
-
-                for agent, resource_id in waiting_agents:
-                    aclient = self.get_agent_client(env.id, agent)
-                    if aclient is not None:
-                        await aclient.resource_event(env.id, agent, resource_id, send_events, status, change, changes)
 
         return 200
